chore: re-implement topic creation
This new implementation allows for multiple topics to be created and worked on
in parallel and utilizes the existing `<Topic>` component for consistent styling. It
could be argued that there's still some room for improvement in the "editing" design
for the `<Topic>` component to make it more clear it's editable. Totally open to
ideas on that.