refactor(api): change how the api is consumed and utilized
Instead of making all the core api methods proxy to the dev methods
depending on the URL, this change utilizes a new `development` prop
to flag whether `api.dev` should be used or not. It also defaults the
`name` / `list` parameter to `todo` and moves the `name` prop from
the `Wrapper` to state as if we ever do support multiple lists it would
probably be changed by a dropdown somewhere within the app.